requirements.txt
imageio==2.31.6
matplotlib==3.7.2
numpy==1.25.2
pandas==2.0.3
Pillow==10.0.0
Pillow==10.3.0
seaborn==0.13.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
